requirements.txt
attrs==20.3.0
Automat==20.2.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
constantly==15.1.0
Flask==1.1.2
hyperlink==20.0.1
idna==2.9
incremental==17.5.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyHamcrest==2.0.2
pyparsing==2.4.7
PyYAML==5.3.1
rdflib==5.0.0
rdflib-jsonld==0.5.0
requests==2.23.0
requests-cache==0.5.2
six==1.14.0
Twisted==20.3.0
urllib3==1.25.9
Werkzeug==1.0.1
zope.interface==5.2.0
